services/web/requirements.txt
Docker==6.1.2
eventlet==0.33.3
Flask==2.3.2
flask-bcrypt==1.0.1
flask-login==0.6.2
flask-socketio==5.3.4
flask-sqlalchemy==3.0.3
flask-wtf==1.1.1
gevent-websocket==0.10.1
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytest==7.3.1